+++ src/elements/Product.php
@@ -341,6 +341,16 @@ class Product extends Element
     /**
      * @inheritdoc
      */
+    public function getCacheTags(): array
+    {
+        return [
+            "productType:$this->typeId",
+        ];
+    }
+
+    /**
+     * @inheritdoc
+     */
     public function getUriFormat()
     {
         $productTypeSiteSettings = $this->getType()->getSiteSettings();
@@ -1100,8 +1110,8 @@ class Product extends Element
     protected static function defineTableAttributes(): array
     {
         return [
-            'id' => ['label' => Craft::t('commerce', 'ID')],
             'title' => ['label' => Craft::t('commerce', 'Title')],
+            'id' => ['label' => Craft::t('commerce', 'ID')],
             'type' => ['label' => Craft::t('commerce', 'Type')],
             'slug' => ['label' => Craft::t('commerce', 'Slug')],
             'uri' => ['label' => Craft::t('commerce', 'URI')],
@@ -1201,6 +1211,11 @@ class Product extends Element
      */
     protected function route()
     {
+        // Make sure that the product is actually live
+        if (!$this->previewing && $this->getStatus() != self::STATUS_LIVE) {
+            return null;
+        }
+
         // Make sure the product type is set to have URLs for this site
         $siteId = Craft::$app->getSites()->currentSite->id;
         $productTypeSiteSettings = $this->getType()->getSiteSettings();