+++ src/services/Carts.php
@@ -44,7 +44,7 @@ class Carts extends Component
      * @var Order
      */
     private $_cart;
-
+    
     /**
      * Get the current cart for this session.
      *
@@ -105,7 +105,8 @@ class Carts extends Component
         $somethingChangedOnTheCart = ($changedIp || $changedOrderLanguage || $changedCustomerId || $changedPaymentCurrency || $changedOrderSiteId);
 
         // If the cart has already been saved (has an ID), then only save if something else changed.
-        if (($this->_cart->id && $somethingChangedOnTheCart) || $forceSave) {
+        // Manual force save only works when the order has not ID
+        if (($this->_cart->id && $somethingChangedOnTheCart) || ($forceSave && !$this->_cart->id)) {
             Craft::$app->getElements()->saveElement($this->_cart, false);
         }