+++ src/services/Emails.php
@@ -329,9 +329,19 @@ class Emails extends Component
             $emailRecord->templatePath = $data['templatePath'];
             $emailRecord->plainTextTemplatePath = $data['plainTextTemplatePath'] ?? null;
             $emailRecord->uid = $emailUid;
-            $emailRecord->pdfId = $pdfUid ? Db::idByUid(Table::PDFS, $pdfUid) : null;
-            $emailRecord->language = $data['language'];
-         
+
+            // todo: remove schema version condition after next beakpoint
+            $projectConfig = Craft::$app->getProjectConfig();
+            $schemaVersion = $projectConfig->get('plugins.commerce.schemaVersion', true);
+
+            if (version_compare($schemaVersion, '3.2.0', '>=')) {
+                $emailRecord->pdfId = $pdfUid ? Db::idByUid(Table::PDFS, $pdfUid) : null;
+            }
+
+            if (version_compare($schemaVersion, '3.2.13', '>=')) {
+                $emailRecord->language = $data['language'] ?? EmailRecord::LOCALE_ORDER_LANGUAGE;
+            }
+
             $emailRecord->save(false);
 
             $transaction->commit();
@@ -863,7 +873,7 @@ class Emails extends Component
      */
     private function _createEmailQuery(): Query
     {
-        return (new Query())
+        $query = (new Query())
             ->select([
                 'emails.id',
                 'emails.name',
@@ -876,12 +886,24 @@ class Emails extends Component
                 'emails.enabled',
                 'emails.templatePath',
                 'emails.plainTextTemplatePath',
-                'emails.pdfId',
-                'emails.language',
                 'emails.uid',
             ])
             ->orderBy('name')
             ->from([Table::EMAILS . ' emails']);
+
+        // todo: remove schema version condition after next beakpoint
+        $projectConfig = Craft::$app->getProjectConfig();
+        $schemaVersion = $projectConfig->get('plugins.commerce.schemaVersion');
+
+        if (version_compare($schemaVersion, '3.2.0', '>=')) {
+            $query->addSelect(['emails.pdfId']);
+        }
+
+        if (version_compare($schemaVersion, '3.2.13', '>=')) {
+            $query->addSelect(['emails.language']);
+        }
+
+        return $query;
     }